requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2020.11.8
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.13
django-filter==22.1
django-silk==5.0.1
djangorestframework==3.13.1
drf-yasg==1.21.3
gprof2dot==2022.7.29
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
